Fix test failed
