+++ components/ws-proxy/pkg/config/config.go
@@ -5,13 +5,19 @@
 package config
 
 import (
+	"context"
 	"encoding/json"
 	"os"
+	"time"
 
 	"golang.org/x/xerrors"
 
+	"github.com/gitpod-io/gitpod/common-go/experiments"
+	"github.com/gitpod-io/gitpod/common-go/log"
 	"github.com/gitpod-io/gitpod/ws-proxy/pkg/proxy"
 )
+
+const experimentsCorsEnabled = "ws_proxy_cors_enabled"
 
 // Config configures this service.
 type Config struct {
@@ -64,5 +70,30 @@ func GetConfig(fn string) (*Config, error) {
 		return nil, xerrors.Errorf("config validation error: %w", err)
 	}
 
+	timeout := time.Minute * 5
+	log.WithField("timeout", timeout).Info("nil CorsEnabled, wait for Feature Flag")
+	experimentsClient := experiments.NewClient(experiments.WithPollInterval(time.Second * 3))
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+	ffValue := waitStringValue(ctx, experimentsClient, experimentsCorsEnabled, "nope", experiments.Attributes{})
+	corsEnabled := ffValue == "true"
+	log.WithField("ffValue", ffValue).WithField("corsEnabled", corsEnabled).Info("feature flag final value")
+
 	return &cfg, nil
+}
+
+func waitStringValue(ctx context.Context, client experiments.Client, experimentName, nopeValue string, attributes experiments.Attributes) string {
+	ticker := time.NewTicker(1 * time.Second)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ctx.Done():
+			return nopeValue
+		case <-ticker.C:
+			value := client.GetStringValue(ctx, experimentName, nopeValue, attributes)
+			if value != nopeValue {
+				return value
+			}
+		}
+	}
 }
