+++ components/ws-proxy/pkg/proxy/routes.go
@@ -34,7 +34,6 @@ import (
 	"github.com/sirupsen/logrus"
 	"golang.org/x/xerrors"
 
-	"github.com/gitpod-io/gitpod/common-go/experiments"
 	"github.com/gitpod-io/gitpod/common-go/log"
 	gitpod "github.com/gitpod-io/gitpod/gitpod-protocol"
 	"github.com/gitpod-io/gitpod/ws-manager/api"
@@ -48,11 +47,7 @@ type RouteHandlerConfig struct {
 	DefaultTransport     http.RoundTripper
 	CorsHandler          mux.MiddlewareFunc
 	WorkspaceAuthHandler mux.MiddlewareFunc
-
-	CorsEnabled bool
 }
-
-const experimentsCorsEnabled = "ws_proxy_cors_enabled"
 
 // RouteHandlerConfigOpt modifies the router handler config.
 type RouteHandlerConfigOpt func(*Config, *RouteHandlerConfig)
@@ -66,23 +61,15 @@ func WithDefaultAuth(infoprov common.WorkspaceInfoProvider) RouteHandlerConfigOp
 
 // NewRouteHandlerConfig creates a new instance.
 func NewRouteHandlerConfig(config *Config, opts ...RouteHandlerConfigOpt) (*RouteHandlerConfig, error) {
-	corsHandler, err := corsHandler(config.GitpodInstallation.Scheme, config.GitpodInstallation.HostName)
+	corsHandler, err := corsHandler(config.CorsEnabled, config.GitpodInstallation.Scheme, config.GitpodInstallation.HostName)
 	if err != nil {
 		return nil, err
 	}
-	experimentsClient := experiments.NewClient(experiments.WithPollInterval(time.Second * 3))
-	ctx, cancel := context.WithDeadline(context.Background(), time.Now().Add(time.Minute*5))
-	defer cancel()
-	ffValue := waitStringValue(ctx, experimentsClient, experimentsCorsEnabled, "nope", experiments.Attributes{})
-	corsEnabled := ffValue == "true"
-	log.WithField("ffValue", ffValue).WithField("corsEnabled", corsEnabled).Info("feature flag final value")
-
 	cfg := &RouteHandlerConfig{
 		Config:               config,
 		DefaultTransport:     createDefaultTransport(config.TransportConfig),
 		CorsHandler:          corsHandler,
 		WorkspaceAuthHandler: func(h http.Handler) http.Handler { return h },
-		CorsEnabled:          corsEnabled,
 	}
 	for _, o := range opts {
 		o(config, cfg)
@@ -175,22 +162,6 @@ type ideRoutes struct {
 	workspaceMustExistHandler mux.MiddlewareFunc
 }
 
-func waitStringValue(ctx context.Context, client experiments.Client, experimentName, nopeValue string, attributes experiments.Attributes) string {
-	ticker := time.NewTicker(1 * time.Second)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return nopeValue
-		case <-ticker.C:
-			value := client.GetStringValue(ctx, experimentName, nopeValue, attributes)
-			if value != nopeValue {
-				return value
-			}
-		}
-	}
-}
-
 func (ir *ideRoutes) HandleSSHHostKeyRoute(route *mux.Route, hostKeyList []ssh.Signer) {
 	shk := make([]struct {
 		Type    string `json:"type"`
@@ -207,9 +178,7 @@ func (ir *ideRoutes) HandleSSHHostKeyRoute(route *mux.Route, hostKeyList []ssh.S
 	}
 	r := route.Subrouter()
 	r.Use(logRouteHandlerHandler("HandleSSHHostKeyRoute"))
-	if ir.Config.CorsEnabled {
-		r.Use(ir.Config.CorsHandler)
-	}
+	r.Use(ir.Config.CorsHandler)
 	r.NewRoute().HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
 		rw.Header().Add("Content-Type", "application/json")
 		rw.Write(byt)
@@ -219,9 +188,8 @@ func (ir *ideRoutes) HandleSSHHostKeyRoute(route *mux.Route, hostKeyList []ssh.S
 func (ir *ideRoutes) HandleCreateKeyRoute(route *mux.Route, hostKeyList []ssh.Signer) {
 	r := route.Subrouter()
 	r.Use(logRouteHandlerHandler("HandleCreateKeyRoute"))
-	if ir.Config.CorsEnabled {
-		r.Use(ir.Config.CorsHandler)
-	}
+	r.Use(ir.Config.CorsHandler)
+
 	r.Use(ir.workspaceMustExistHandler)
 	r.Use(ir.Config.WorkspaceAuthHandler)
 
@@ -285,9 +253,7 @@ func extractCloseErrorCode(errStr string) string {
 func (ir *ideRoutes) HandleSSHOverWebsocketTunnel(route *mux.Route, sshGatewayServer *sshproxy.Server) {
 	r := route.Subrouter()
 	r.Use(logRouteHandlerHandler("HandleSSHOverWebsocketTunnel"))
-	if ir.Config.CorsEnabled {
-		r.Use(ir.Config.CorsHandler)
-	}
+	r.Use(ir.Config.CorsHandler)
 	r.Use(ir.workspaceMustExistHandler)
 	r.Use(ir.Config.WorkspaceAuthHandler)
 
@@ -331,9 +297,7 @@ func (ir *ideRoutes) HandleSSHOverWebsocketTunnel(route *mux.Route, sshGatewaySe
 func (ir *ideRoutes) HandleDirectSupervisorRoute(route *mux.Route, authenticated bool) {
 	r := route.Subrouter()
 	r.Use(logRouteHandlerHandler(fmt.Sprintf("HandleDirectSupervisorRoute (authenticated: %v)", authenticated)))
-	if ir.Config.CorsEnabled {
-		r.Use(ir.Config.CorsHandler)
-	}
+	r.Use(ir.Config.CorsHandler)
 	r.Use(ir.workspaceMustExistHandler)
 	if authenticated {
 		r.Use(ir.Config.WorkspaceAuthHandler)
@@ -418,9 +382,7 @@ type BlobserveInlineVars struct {
 func (ir *ideRoutes) HandleRoot(route *mux.Route) {
 	r := route.Subrouter()
 	r.Use(logRouteHandlerHandler("handleRoot"))
-	if ir.Config.CorsEnabled {
-		r.Use(ir.Config.CorsHandler)
-	}
+	r.Use(ir.Config.CorsHandler)
 	r.Use(ir.workspaceMustExistHandler)
 
 	proxyPassWoSensitiveCookies := sensitiveCookieHandler(ir.Config.Config.GitpodInstallation.HostName)(proxyPass(ir.Config, ir.InfoProvider, workspacePodResolver))
@@ -554,9 +516,7 @@ func installDebugWorkspaceRoutes(r *mux.Router, config *RouteHandlerConfig, info
 	}
 
 	r.Use(logHandler)
-	if config.CorsEnabled {
-		r.Use(config.CorsHandler)
-	}
+	r.Use(config.CorsHandler)
 	r.Use(config.WorkspaceAuthHandler)
 	// filter all session cookies
 	r.Use(sensitiveCookieHandler(config.Config.GitpodInstallation.HostName))
@@ -695,7 +655,13 @@ func buildWorkspacePodURL(protocol api.PortProtocol, ipAddress string, port stri
 }
 
 // corsHandler produces the CORS handler for workspaces.
-func corsHandler(scheme, hostname string) (mux.MiddlewareFunc, error) {
+func corsHandler(enabled bool, scheme, hostname string) (mux.MiddlewareFunc, error) {
+	if !enabled {
+		// empty handler
+		return func(h http.Handler) http.Handler {
+			return h
+		}, nil
+	}
 	origin := fmt.Sprintf("%s://%s", scheme, hostname)
 
 	domainRegex := strings.ReplaceAll(hostname, ".", "\\.")
