+++ components/ws-proxy/pkg/proxy/routes_test.go
@@ -95,6 +95,7 @@ var (
 		BuiltinPages: BuiltinPagesConfig{
 			Location: "../../public",
 		},
+		CorsEnabled: false,
 	}
 )
